+++ docs/package.json
@@ -17,7 +17,7 @@
     "copy-webpack-plugin": "^0.2.0",
     "es6-promise": "^3.0.2",
     "extract-text-webpack-plugin": "^0.9.1",
-    "gatsby": "^0.7.2",
+    "gatsby": "^0.7.3",
     "graphiql": "^0.4.2",
     "graphql": "^0.4.13",
     "jeet": "^6.1.2",
